[jit] fix bug in CompilationUnit::define
We were re-ordering definitions to put `__init__` first, but not
reordering resolvers. In reality, resolvers are always the same so this
doesn't matter. I may make `define()` take a single resolver in a future
PR.

Either way, this change is good because it avoids a copy of the whole
AST during `define()` as well.

gh-metadata: pytorch pytorch 21886 gh/suo/64/head